+++ packages/ui5-app/webapp/test/e2e/conf.js
@@ -1,0 +1,9 @@
+exports.config = {
+    profile: "integration",
+    baseUrl: "http://localhost:1081/index.html",
+    browsers: [
+        {
+            browserName: process.env.HEADLESS ? "chromeHeadless" : "chrome"
+        }
+    ]
+};