+++ packages/ui5-middleware-livereload/lib/livereload.js
@@ -2,6 +2,26 @@ const connectLivereload = require("connect-livereload");
 const livereload = require("livereload");
 const path = require("path");
 const log = require("@ui5/logger").getLogger("server:custommiddleware:livereload");
+const portfinder = require("portfinder"); 
+
+/**
+ * Parses the configuration option. If the port passed then it returns with it.
+ * If not passed it returns with the following free port after the deafult port.
+ * @param {Object} options the entered config option
+ * @param {number} defaultPort the port which is defaulted
+ * @returns {number}
+ */
+const getPortForLivereload = async (options, defaultPort) => {
+    if (options.configuration && options.configuration.port) {
+        return options.configuration.port;
+    }
+    try {
+        portfinder.basePort = defaultPort;
+        return await portfinder.getPortPromise();
+    } catch {
+        return defaultPort;
+    }
+}
 
 /**
  * Custom UI5 Server middleware example
@@ -18,11 +38,8 @@ const log = require("@ui5/logger").getLogger("server:custommiddleware:livereload
  * @param {string} [parameters.options.configuration] Custom server middleware configuration if given in ui5.yaml
  * @returns {function} Middleware function to use
  */
-module.exports = ({ resources, options }) => {
-    let port = 35729;
-    if (options.configuration && options.configuration.port) {
-        port = options.configuration.port;
-    }
+module.exports = async ({ resources, options }) => {
+    let port = await getPortForLivereload(options, 35729);
     let watchPath = "webapp";
     // due to compatibility reasons we keep the path as watchPath (watchPath has higher precedence than path)
     if (options.configuration && (options.configuration.watchPath || options.configuration.path)) {
@@ -32,22 +49,33 @@ module.exports = ({ resources, options }) => {
     if (options.configuration && options.configuration.extraExts) {
         extraExts = options.configuration.extraExts;
     }
-    let debug = true;
+    let debug = false;
     if (options.configuration && options.configuration.debug) {
         debug = options.configuration.debug;
     }
     const livereloadServer = livereload.createServer(
         {
             debug: debug,
-            extraExts: extraExts,
+            extraExts: extraExts ? extraExts.split(",") : undefined,
             port: port
         },
         () => {
             log.info("Livereload server started!");
         }
     );
-    debug ? log.info(`Livereload connecting to port ${port} for path ${watchPath}`) : null;
-    livereloadServer.watch(path.join(process.cwd(), watchPath));
+    
+	if (Array.isArray(watchPath)) {
+		let watchPaths = [];
+		for (let i = 0; i < watchPath.length; i++) {
+			watchPaths.push(path.join(process.cwd(), watchPath[i]));
+		}
+		debug ? log.info(`Livereload connecting to port ${port} for paths ${watchPaths}`) : null;
+		livereloadServer.watch(watchPaths);
+	} else {
+		debug ? log.info(`Livereload connecting to port ${port} for path ${watchPath}`) : null;
+		livereloadServer.watch(path.join(process.cwd(), watchPath));
+	}
+	
     // connect-livereload already holds the 
     // method sig (req, res, next)
     return connectLivereload({