+++ packages/ui5-middleware-stringreplacer/lib/stringreplacer.js
@@ -1,0 +1,148 @@
+const log = require("@ui5/logger").getLogger("server:custommiddleware:stringreplacer");
+
+const minimatch = require("minimatch");
+const etag = require("etag");
+const fresh = require("fresh");
+
+function isFresh(req, res) {
+	return fresh(req.headers, {
+		"etag": res.getHeader("ETag")
+	});
+}
+
+// TODO: for now, we duplicate the code of the ui5-task-stringreplacer and the ui5-middleware-stringreplacer
+//       BUT for the future we should consider a reuse module for the middlewares and tasks in this repository.
+
+// BEGIN: copy of code from ui5-task-stringreplacer
+
+const dotenv = require("dotenv");
+const escapeRegExp = require('lodash.escaperegexp');
+const replaceStream = require("replacestream");
+
+dotenv.config();
+
+// get all environment variables
+const envVariables = process.env;
+
+// manage placeholders
+let placeholderStrings = {};
+function addPlaceholderString(key, value) {
+  placeholderStrings[key] = value;
+}
+
+// loop through env variables to find keys which are having prefix 'stringreplacer'
+if (typeof envVariables === "object") {
+  for (key in envVariables) {
+    // env variable should start with 'stringreplacer' and should in format 'stringreplacer.placeholder'
+    if (/^stringreplacer\.(.+)$/i.test(key)) {
+      let placeholderString = /^stringreplacer\.(.+)$/i.exec(key)[1];
+      addPlaceholderString(placeholderString, envVariables[key]);
+    }
+  }
+}
+
+// create the helper function to pipe the stream and replace the placeholders
+function createReplacePlaceholdersDestination({resource, isDebug}) {
+  const replaceStreamRegExp = `(${Object.keys(placeholderStrings).map(placeholder => {
+    return escapeRegExp(placeholder);
+  }).join("|")})`;
+  return replaceStream(new RegExp(replaceStreamRegExp, "g"), (match) => {
+    isDebug && log.info(`${resource.getPath()} matched: ${match}; replacing with ${placeholderStrings[match]}`);
+    return placeholderStrings[match];
+  });
+}
+
+// END: copy of code from ui5-task-stringreplacer
+
+
+/**
+ * Custom UI5 Server middleware example
+ *
+ * @param {Object} parameters Parameters
+ * @param {Object} parameters.resources Resource collections
+ * @param {module:@ui5/fs.AbstractReader} parameters.resources.all Reader or Collection to read resources of the
+ *                                        root project and its dependencies
+ * @param {module:@ui5/fs.AbstractReader} parameters.resources.rootProject Reader or Collection to read resources of
+ *                                        the project the server is started in
+ * @param {module:@ui5/fs.AbstractReader} parameters.resources.dependencies Reader or Collection to read resources of
+ *                                        the projects dependencies
+ * @param {Object} parameters.options Options
+ * @param {string} [parameters.options.configuration] Custom server middleware configuration if given in ui5.yaml
+ * @param {object} parameters.middlewareUtil Specification version dependent interface to a
+ *                                        [MiddlewareUtil]{@link module:@ui5/server.middleware.MiddlewareUtil} instance
+ * @returns {function} Middleware function to use
+ */
+module.exports = function createMiddleware({ resources, options, middlewareUtil }) {
+
+  const isDebug = options.configuration && options.configuration.debug;
+
+  let filesToInclude = [];
+  if (options.configuration) {
+    // extract the configuration of files to be included
+    const { files } = options.configuration;
+    if (Array.isArray(files)) {
+      filesToInclude = [...files];
+    } else {
+      filesToInclude.push(files);
+    }
+    // extract the placeholder strings from the configuration
+    const { replace } = options.configuration;
+    replace && replace.forEach(entry => {
+      addPlaceholderString(entry.placeholder, entry.value);
+    });
+  }
+
+  // helper to determine whether to handle the request or not
+  function handleRequest(path) {
+    return filesToInclude.some(value => {
+      return minimatch(path, value);
+    });
+  }
+
+  // returns the middleware function
+  return async function stringreplacer(req, res, next) {
+
+    const pathname = middlewareUtil.getPathname(req);
+    const resource = await resources.all.byPath(pathname);
+    if (!resource) { // Resource not found
+      next();
+      return;
+    }
+
+    if (handleRequest(req.path)) {
+      //isDebug && log.info(`handling ${req.path}`);
+
+      // enable ETag caching
+      res.setHeader("ETag", etag(resource.getStatInfo()));
+
+      if (isFresh(req, res)) {
+        // client has a fresh copy of the resource
+        res.statusCode = 304;
+        res.end();
+        return;
+      }
+
+      // determine charset and content-type
+      const {contentType, charset} = middlewareUtil.getMimeInfo(resource.getPath());
+      if (!res.getHeader("Content-Type")) {
+        res.setHeader("Content-Type", contentType);
+      }
+
+      // stream replacement only works for UTF-8 resources!
+      let stream = resource.getStream();
+      if (!charset || charset === "UTF-8") {
+        stream.setEncoding("utf8");
+        stream = stream.pipe(createReplacePlaceholdersDestination({resource, isDebug}));
+      } else {
+        isDebug && log.warn(`skipping placeholder replacement for non-UTF-8 resource ${req.path}`);
+      }
+      stream.pipe(res);
+
+      return;
+    } else {
+      next();
+      return;
+    }
+
+  }
+}