+++ packages/ui5-app/CHANGELOG.md
@@ -3,6 +3,153 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.3](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.8.2...ui5-app@0.8.3) (2020-11-12)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+## [0.8.2](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.8.1...ui5-app@0.8.2) (2020-11-03)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+## [0.8.1](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.8.0...ui5-app@0.8.1) (2020-09-29)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+# [0.8.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.7.2...ui5-app@0.8.0) (2020-09-29)
+
+
+### Features
+
+* **simpleproxy:** allow to increase payload limit ([7c5bbfb](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/7c5bbfb72d2196b02bc14e7bd7313de0584b5e7e))
+* **stringreplacer:** new middleware and alignment of task ([40dcb4d](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/40dcb4d4442b0262699a779a13b565d8bba07a87))
+
+
+
+
+
+## [0.7.2](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.7.1...ui5-app@0.7.2) (2020-09-20)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+## [0.7.1](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.7.0...ui5-app@0.7.1) (2020-08-03)
+
+
+### Bug Fixes
+
+* use absolute path for servestatic middleware ([c6086ec](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/c6086ec057411188b240e722e751a0341c70647a)), closes [#278](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/278)
+
+
+
+
+
+# [0.7.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.6.2...ui5-app@0.7.0) (2020-07-17)
+
+
+### Features
+
+* Enhance simpleproxy with adding configuration option for http headers ([#250](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/250)) ([8cfd528](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/8cfd528db846e40d940f854adc215a18c6b002ae))
+
+
+
+
+
+## [0.6.2](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.6.1...ui5-app@0.6.2) (2020-07-06)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+## [0.6.1](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.6.0...ui5-app@0.6.1) (2020-06-23)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+# [0.6.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.5.0...ui5-app@0.6.0) (2020-05-27)
+
+
+### Features
+
+* new custom task - stringsreplacer ([#215](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/215)) ([2e02864](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/2e02864ce89da0f200c3a7da440706370771a8a7))
+
+
+
+
+
+# [0.5.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.4.0...ui5-app@0.5.0) (2020-05-24)
+
+
+### Bug Fixes
+
+* **docu:** extra extensions config for livereload ([1f6149a](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/1f6149a97354a3360c608ab2027086f174cd908e))
+
+
+### Features
+
+* **pwa:** added pwa build for sample application ([c36baf2](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/c36baf24ed93e4e3634374c7ddcd426b8818876f))
+
+
+
+
+
+# [0.4.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.3.2...ui5-app@0.4.0) (2020-05-06)
+
+
+### Features
+
+* **zipper:** allow to include additional files ([#198](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/198)) ([abf6d01](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/abf6d016d955a49763c59ac508b6267a760052e9))
+
+
+
+
+
+## [0.3.2](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.3.1...ui5-app@0.3.2) (2020-05-05)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+## [0.3.1](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.3.0...ui5-app@0.3.1) (2020-05-02)
+
+**Note:** Version bump only for package ui5-app
+
+
+
+
+
+# [0.3.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.2.0...ui5-app@0.3.0) (2020-05-01)
+
+
+### Features
+
+* **i18n:** ui5-task-i18ncheck ([#183](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/183)) ([5b4b033](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/5b4b033f8b7bdd57f9d1a93045740b55747b1611))
+
+
+
+
+
 # [0.2.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-app@0.1.10...ui5-app@0.2.0) (2020-04-04)