+++ packages/ui5-middleware-livereload/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.3](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livereload@0.4.2...ui5-middleware-livereload@0.4.3) (2020-09-29)
+
+
+### Bug Fixes
+
+* **livereload:** ensure proper concat for extra exts ([#311](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/311)) ([cdf4b63](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/cdf4b633d9b4ad30e209038e08dcfe24277712cc))
+
+
+
+
+
+## [0.4.2](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livereload@0.4.1...ui5-middleware-livereload@0.4.2) (2020-09-20)
+
+**Note:** Version bump only for package ui5-middleware-livereload
+
+
+
+
+
+## [0.4.1](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livereload@0.4.0...ui5-middleware-livereload@0.4.1) (2020-08-03)
+
+**Note:** Version bump only for package ui5-middleware-livereload
+
+
+
+
+
+# [0.4.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livereload@0.3.0...ui5-middleware-livereload@0.4.0) (2020-07-17)
+
+
+### Bug Fixes
+
+* default to debug false ([#253](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/253)) ([0926ae4](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/0926ae4e2235766ef744333d13064e9e948f28b6))
+
+
+### Features
+
+* ui5-middleware-livereload - watch multiple paths ([#259](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/259)) ([16005f6](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/16005f6373db4ed2ead127fb4cf009465c738cc4))
+
+
+
+
+
+# [0.3.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livereload@0.2.0...ui5-middleware-livereload@0.3.0) (2020-05-24)
+
+
+### Bug Fixes
+
+* **docu:** extra extensions config for livereload ([1f6149a](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/1f6149a97354a3360c608ab2027086f174cd908e))
+
+
+### Features
+
+* **pwa:** added pwa build for sample application ([c36baf2](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/c36baf24ed93e4e3634374c7ddcd426b8818876f))
+
+
+
+
+
+# [0.2.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livereload@0.1.10...ui5-middleware-livereload@0.2.0) (2020-05-01)
+
+
+### Features
+
+* Automatically set a non-taken port for livereload server ([#195](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/195)) ([9ca6037](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/9ca6037a8977e3ed5279d834bb7347a6cc6f0eed))
+
+
+
+
+
 ## [0.1.10](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livereload@0.1.9...ui5-middleware-livereload@0.1.10) (2020-04-04)
 
 **Note:** Version bump only for package ui5-middleware-livereload