+++ packages/ui5-middleware-livetranspile/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livetranspile@0.2.3...ui5-middleware-livetranspile@0.2.4) (2020-09-20)
+
+**Note:** Version bump only for package ui5-middleware-livetranspile
+
+
+
+
+
+## [0.2.3](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livetranspile@0.2.2...ui5-middleware-livetranspile@0.2.3) (2020-08-03)
+
+**Note:** Version bump only for package ui5-middleware-livetranspile
+
+
+
+
+
+## [0.2.2](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livetranspile@0.2.1...ui5-middleware-livetranspile@0.2.2) (2020-07-17)
+
+**Note:** Version bump only for package ui5-middleware-livetranspile
+
+
+
+
+
+## [0.2.1](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livetranspile@0.2.0...ui5-middleware-livetranspile@0.2.1) (2020-05-27)
+
+
+### Bug Fixes
+
+* safeguard options (closes [#217](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/217)) ([#218](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/218)) ([bcea7c9](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/bcea7c933a65764aa374397fdb1366cec3ed208f))
+
+
+
+
+
+# [0.2.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livetranspile@0.1.9...ui5-middleware-livetranspile@0.2.0) (2020-05-24)
+
+
+### Features
+
+* **pwa:** added pwa build for sample application ([c36baf2](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/c36baf24ed93e4e3634374c7ddcd426b8818876f))
+
+
+
+
+
+## [0.1.9](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livetranspile@0.1.8...ui5-middleware-livetranspile@0.1.9) (2020-05-01)
+
+**Note:** Version bump only for package ui5-middleware-livetranspile
+
+
+
+
+
 ## [0.1.8](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-livetranspile@0.1.7...ui5-middleware-livetranspile@0.1.8) (2020-04-04)