+++ packages/ui5-middleware-servestatic/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-servestatic@0.3.2...ui5-middleware-servestatic@0.3.3) (2020-09-20)
+
+**Note:** Version bump only for package ui5-middleware-servestatic
+
+
+
+
+
+## [0.3.2](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-servestatic@0.3.1...ui5-middleware-servestatic@0.3.2) (2020-08-03)
+
+
+### Bug Fixes
+
+* use absolute path for servestatic middleware ([c6086ec](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/c6086ec057411188b240e722e751a0341c70647a)), closes [#278](https://github.com/petermuessig/ui5-ecosystem-showcase/issues/278)
+
+
+
+
+
+## [0.3.1](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-servestatic@0.3.0...ui5-middleware-servestatic@0.3.1) (2020-07-17)
+
+**Note:** Version bump only for package ui5-middleware-servestatic
+
+
+
+
+
+# [0.3.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-servestatic@0.2.0...ui5-middleware-servestatic@0.3.0) (2020-05-24)
+
+
+### Features
+
+* **pwa:** added pwa build for sample application ([c36baf2](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/c36baf24ed93e4e3634374c7ddcd426b8818876f))
+
+
+
+
+
+# [0.2.0](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-servestatic@0.1.6...ui5-middleware-servestatic@0.2.0) (2020-05-01)
+
+
+### Features
+
+* add error handling ([1f51951](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/1f519513f7989f6640f593666cfe9d45f68643a1))
+* add option to allow paths from .env file ([6152658](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/615265803985c6ed8fd9f6bf785a9eccc42e6fb2))
+* use regex logic to detect .env config option ([d7b0c67](https://github.com/petermuessig/ui5-ecosystem-showcase/commit/d7b0c6713343ae80f7115dd7b5cce8382dfb64d6))
+
+
+
+
+
 ## [0.1.6](https://github.com/petermuessig/ui5-ecosystem-showcase/compare/ui5-middleware-servestatic@0.1.5...ui5-middleware-servestatic@0.1.6) (2020-04-04)
 
 **Note:** Version bump only for package ui5-middleware-servestatic